+++ spi_sw.c
@@ -6,7 +6,7 @@ void spi_init(struct spi_t *spi, pin_t sdi, pin_t clk) {
 }
 
 void spi_write(struct spi_t *spi, uint8_t byte) {
-    for (int i = 0; i < 8; ++i) {
+    for (uint8_t i = 0; i < 8; ++i) {
         if (byte & (1 << 7)) {
             pin_write(spi->sdi, true);
         } else {
